Replace synthetic data with real historical data for NFL and NCAAF
Real, no-API-key data sources, discovered and validated against this
session's actual (restrictive) network policy:

- NFL: data_pipeline/sources/nfl_nflverse.py pulls real play-by-play
  from nflverse-data's public GitHub release assets (epa, success,
  down/distance, and — verified against real blowout games for sign
  convention — real closing spread_line/total_line). 9 seasons
  (2015-2023), 2,458 real games ingested.
- NCAAF: data_pipeline/sources/ncaaf_cfbfastr.py parses a local
  checkout of sportsdataverse/cfbfastr-data (schedules + play-by-play,
  FBS-vs-FBS only) for the same, including its own embedded real
  closing spread/total. 10 seasons (2012-2021), 7,006 real games.
- scripts/ingest_real_data.py orchestrates both, replacing
  scripts/generate_sample_data.py as the primary path (kept for
  pipeline-mechanics testing only).
- NBA: data_pipeline/sources/nba_nba_api.py is real, correct nba_api
  code, explicitly NOT executed here — stats.nba.com was unreachable
  from this session's network policy. Documented clearly rather than
  faked.
- data_pipeline/config.py: ODDS_API_KEY now loads from a local .env
  (python-dotenv, already-gitignored) or Streamlit Cloud's secrets
  store, in addition to a real env var.

Ran the real walk-forward backtest (models/production.py) against both
sports and recorded the honest result in README: the plain Elo
baseline currently beats the GBM out-of-sample on both (NFL Brier
0.232 vs 0.240; NCAAF 0.187 vs 0.193) — reported plainly rather than
buried, per this project's own "beat Elo or the complexity isn't
earning its keep" standard.

Also fixes two real bugs surfaced while validating against real data:
- models/gbm_model.py::evaluate_classifier crashed on real data because
  pandas' nullable Int64 dtype (home_win's dtype) converts to an
  object-dtype array via np.asarray(), which broke sklearn's target-
  type inference. Synthetic data happened not to trigger it.
- requirements.txt: pip installing nfl_data_py (only used transiently
  to confirm its source URLs, never as a runtime dependency) silently
  downgraded pandas to <2.0 and broke rolling-feature alignment with
  no error — documented prominently so nobody repeats it.

+++ models/gbm_model.py
@@ -49,8 +49,13 @@ def calibration_table(y_true: np.ndarray, y_prob: np.ndarray, n_buckets: int = 1
 
 
 def evaluate_classifier(y_true: np.ndarray, y_prob: np.ndarray) -> CalibrationReport:
-    y_true = np.asarray(y_true)
-    y_prob = np.clip(np.asarray(y_prob), 1e-6, 1 - 1e-6)
+    # np.asarray() on a pandas nullable Int64 Series (which home_win is,
+    # from build_feature_frame) yields an OBJECT-dtype array, not int64 —
+    # sklearn's type_of_target then can't tell it's binary and raises. Force
+    # a concrete numeric dtype so this works regardless of the pandas dtype
+    # the caller happened to hand in.
+    y_true = np.asarray(y_true, dtype=np.float64)
+    y_prob = np.clip(np.asarray(y_prob, dtype=np.float64), 1e-6, 1 - 1e-6)
     return CalibrationReport(
         brier_score=float(brier_score_loss(y_true, y_prob)),
         log_loss=float(log_loss(y_true, y_prob)),
